released 5.4
fix #226
fix #227
fix #228
natively support non-blocking `FILE*` input, i.e. for `fcntl` `O_NONBLOCK` the regex matchers will wait for input to become available again instead of giving up with an error; changed `reflex::Input::Handler` see documentation

+++ examples/lua.hpp
@@ -112,7 +112,7 @@ class Transpiler {
   // integer literal expression
   struct Integer : public Expression {
     Integer(IntegerLiteral n) : literal(n) { }
-    virtual void transpile(Transpiler& transpiler) { fprintf(transpiler.out, "%llu", literal); }
+    virtual void transpile(Transpiler& transpiler) { fprintf(transpiler.out, "%llu", static_cast<unsigned long long>(literal)); }
     IntegerLiteral literal;
   };
